+++ .github/workflows/test_accuracy.yml
@@ -25,7 +25,7 @@ jobs:
           uv sync --locked --extra tests --extra-index-url https://download.pytorch.org/whl/cpu
       - name: Prepare test data
         run: |
-          uv run python tests/accuracy/prepare_data.py -d data
+          uv run python tests/accuracy/download_models.py -d data -j tests/accuracy/public_scope.json -l
       - name: Run Python Test
         run: |
           uv run pytest --data=./data tests/accuracy/test_accuracy.py
